+++ lib/animation.dart
@@ -2,31 +2,38 @@ import 'package:tremble/sprite.dart';
 import 'package:tremble/tex_area.dart';
 
 enum AnimMode {
-  playStop,
-  playReset,
+  playOnce,
+  playOnceReset,
   pingPong,
   loop,
 }
 
-class AnimationData {
+int _clampToRange(int value, int min, int max) {
+  if (value < min) return min;
+  if (value > max) return max;
+  return value;
+}
+
+class AnimationData<T extends Enum> {
   AnimationData({
     required this.name,
     required this.frames,
     required this.speed,
-  });
+  })  : assert(frames.isNotEmpty, 'AnimationData "$name" must contain at least 1 frame'),
+        assert(speed >= 0, 'speed must be non-negative');
 
-  final String name;
+  final T name;
   final List<TexArea> frames;
   final double speed;
 }
 
-class Animation extends Sprite {
+class Animation<T extends Enum> extends Sprite {
   Animation({
-    required List<AnimationData> animations,
+    required List<AnimationData<T>> animations,
     required super.position,
     int index = 0,
     this.mode = AnimMode.loop,
-    this.reverse = false,
+    bool reverse = false,
     super.originX = 0.5,
     super.originY = 0.5,
     super.opacity = 255,
@@ -35,22 +42,38 @@ class Animation extends Sprite {
   })  : assert(animations.isNotEmpty, "you have to provide atleast 1 AnimationData"),
         _animations = Map.fromEntries(animations.map((e) => MapEntry(e.name, e))),
         _state = animations.first.name,
+        _reverse = reverse,
         _direction = reverse ? -1 : 1,
         _index = reverse ? animations.first.frames.length - 1 : index,
         _timer = reverse ? (animations.first.frames.length - 1).toDouble() : index.toDouble(),
         super(
-            texture: animations.first.frames[reverse ? animations.first.frames.length - 1 : index]);
+          texture: animations.first.frames[reverse ? animations.first.frames.length - 1 : index],
+        ) {
+    assert(
+      reverse || (index >= 0 && index < animations.first.frames.length),
+      "initial index is out of bounds for the first animation",
+    );
+  }
 
-  final Map<String, AnimationData> _animations;
+  final Map<T, AnimationData<T>> _animations;
 
   bool paused = false;
   AnimMode mode;
-  bool reverse;
-  String _state;
+
+  bool _reverse;
+  bool get reverse => _reverse;
+  set reverse(bool value) {
+    _reverse = value;
+    _direction = value ? -1 : 1;
+  }
+
+  T _state;
+  T get currentState => _state;
+
   double _timer;
   int _direction;
 
-  AnimationData get currentAnimation => _animations[_state]!;
+  AnimationData<T> get currentAnimation => _animations[_state]!;
 
   bool _finished = false;
   bool get finished => _finished;
@@ -58,43 +81,51 @@ class Animation extends Sprite {
   int _index;
   int get index => _index;
   set index(int val) {
-    _index = val;
-    _timer = val.toDouble();
-  }
-
-  int _actualIndex(int frameCount) {
-    return reverse ? (frameCount - 1 - _index) : _index;
+    final frameCount = currentAnimation.frames.length;
+    _index = _clampToRange(val, 0, frameCount - 1);
+    _timer = _index.toDouble();
   }
 
   void update(double deltaTime) {
-    if (paused) {
-      if (_index >= 0 && _index < currentAnimation.frames.length) {
-        texture = currentAnimation.frames[_actualIndex(currentAnimation.frames.length)];
-      }
-      return;
-    }
+    if (paused) return;
 
     _finished = false;
 
-    if (mode == AnimMode.playStop) {
-      _updatePlayStop(deltaTime);
-    } else if (mode == AnimMode.playReset) {
-      _updatePlayReset(deltaTime);
-    } else if (mode == AnimMode.loop) {
-      _updateLoop(deltaTime);
-    } else if (mode == AnimMode.pingPong) {
-      _updatePingPong(deltaTime);
+    switch (mode) {
+      case AnimMode.playOnce:
+        _updatePlayOnce(deltaTime);
+        break;
+      case AnimMode.playOnceReset:
+        _updatePlayOnceReset(deltaTime);
+        break;
+      case AnimMode.loop:
+        _updateLoop(deltaTime);
+        break;
+      case AnimMode.pingPong:
+        _updatePingPong(deltaTime);
+        break;
     }
   }
 
-  void _updatePlayStop(double deltaTime) {
-    final frameCount = currentAnimation.frames.length;
-    _timer += _direction * currentAnimation.speed * deltaTime;
+  void _updatePlayOnce(double deltaTime) {
+    final anim = currentAnimation;
+
+    final frameCount = anim.frames.length;
+    if (frameCount <= 1) {
+      _index = 0;
+      _finished = true;
+      paused = true;
+      texture = anim.frames[_index];
+      return;
+    }
+
+    _timer += _direction * anim.speed * deltaTime;
 
     if (_direction > 0) {
       if (_timer >= frameCount - 1) {
         _index = frameCount - 1;
         _finished = true;
+        paused = true;
       } else {
         _index = _timer.toInt();
       }
@@ -102,23 +133,34 @@ class Animation extends Sprite {
       if (_timer < 0) {
         _index = 0;
         _finished = true;
+        paused = true;
       } else {
         _index = _timer.toInt();
       }
     }
 
-    texture = currentAnimation.frames[_actualIndex(frameCount)];
+    texture = anim.frames[_index];
   }
 
-  void _updatePlayReset(double deltaTime) {
-    final frameCount = currentAnimation.frames.length;
-    _timer += _direction * currentAnimation.speed * deltaTime;
+  void _updatePlayOnceReset(double deltaTime) {
+    final anim = currentAnimation;
+    final frameCount = anim.frames.length;
+    if (frameCount <= 1) {
+      _index = 0;
+      _finished = true;
+      paused = true;
+      texture = anim.frames[_index];
+      return;
+    }
+
+    _timer += _direction * anim.speed * deltaTime;
 
     if (_direction > 0) {
       if (_timer >= frameCount) {
         _index = 0;
         _timer = 0;
         _finished = true;
+        paused = true;
       } else {
         _index = _timer.toInt();
       }
@@ -127,66 +169,71 @@ class Animation extends Sprite {
         _index = frameCount - 1;
         _timer = (frameCount - 1).toDouble();
         _finished = true;
+        paused = true;
       } else {
         _index = _timer.toInt();
       }
     }
 
-    texture = currentAnimation.frames[_actualIndex(frameCount)];
+    texture = anim.frames[_index];
   }
 
   void _updateLoop(double deltaTime) {
-    final frameCount = currentAnimation.frames.length;
-    _timer += _direction * currentAnimation.speed * deltaTime;
+    final anim = currentAnimation;
 
-    if (_direction > 0) {
-      if (_timer >= frameCount) {
-        _timer -= frameCount;
-      }
-    } else {
-      if (_timer < 0) {
-        _timer += frameCount;
-      }
+    final frameCount = anim.frames.length;
+    if (frameCount <= 1) {
+      _index = 0;
+      texture = anim.frames[_index];
+      return;
     }
 
-    _index = _timer.toInt();
-    texture = currentAnimation.frames[_actualIndex(frameCount)];
+    _timer += _direction * anim.speed * deltaTime;
+    _timer %= frameCount;
+
+    _index = _clampToRange(_timer.toInt(), 0, frameCount - 1);
+    texture = anim.frames[_index];
   }
 
   void _updatePingPong(double deltaTime) {
-    final frameCount = currentAnimation.frames.length;
-    _timer += _direction * currentAnimation.speed * deltaTime;
+    final anim = currentAnimation;
 
-    if (_direction > 0) {
-      if (_timer >= frameCount - 1) {
-        _timer = (frameCount - 1) - (_timer - (frameCount - 1));
-        _direction = -1;
-      }
-    } else {
-      if (_timer < 0) {
-        _timer = -_timer;
-        _direction = 1;
-      }
+    final frameCount = anim.frames.length;
+    if (frameCount <= 1) {
+      _index = 0;
+      texture = anim.frames[_index];
+      return;
     }
 
-    _index = _timer.toInt();
-    texture = currentAnimation.frames[_actualIndex(frameCount)];
+    final cycleLength = 2 * (frameCount - 1);
+    _timer += _direction * anim.speed * deltaTime;
+
+    _timer %= cycleLength;
+
+    final folded = _timer <= frameCount - 1 ? _timer : cycleLength - _timer;
+
+    _index = _clampToRange(folded.toInt(), 0, frameCount - 1);
+    texture = anim.frames[_index];
   }
 
-  void setAnimation(String name, {int? fromFrame}) {
+  void setAnimation(T name, {int? fromFrame}) {
     if (_state == name) return;
     resetAnimation(name, fromFrame: fromFrame);
   }
 
-  void resetAnimation(String name, {int? fromFrame}) {
+  void resetAnimation(T name, {int? fromFrame}) {
+    assert(_animations.containsKey(name), 'Animation "$name" was not registered');
+
     paused = false;
     _state = name;
     _direction = reverse ? -1 : 1;
 
+    final frameCount = currentAnimation.frames.length;
+
     if (fromFrame != null) {
-      index = fromFrame;
+      index = _clampToRange(fromFrame, 0, frameCount - 1);
     } else if (reverse) {
-      index = currentAnimation.frames.length - 1;
+      index = frameCount - 1;
     } else {
       index = 0;
     }